requirements.txt
langchain_community==0.3.7
langchain_core==0.3.19
langchain_groq==0.2.1
PyPDF2==3.0.1
python-dotenv==1.0.1
python_docx==1.1.2
streamlit==1.39.0
